handle click on logo link (reset form filters and page)
